requirements.txt
google-cloud-bigquery==3.10.0
google-cloud-bigquery[pandas]==3.10.0
pretty_html_table==0.9.8
sendgrid==6.10.0
pandas-gbq==0.13.1
